requirements.txt
opencv-python
mmdet
mmcv
numpy
Pillow
imagehash
sort-tracker
pokemontcgsdk
